change review page path
- change review page path
- revert yarn.lock
- remove package-lock.json
